Fix a bunch of gcc-3.3 related warnings
git-svn-id: http://svn.ratbox.org/svnroot/trunk@13951 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ adns/general.c
@@ -220,7 +220,7 @@ adns_status adns_rr_info(adns_rrtype type,
   st= typei->convstring(&vb,datap);
   if (st) goto x_freevb;
   if (!adns__vbuf_append(&vb,(const byte *)"",1)) { st= adns_s_nomemory; goto x_freevb; }
-  assert(strlen((const char *)vb.buf) == vb.used-1);
+  assert(strlen((const char *)vb.buf) == (size_t)vb.used-1);
   *data_r= MyRealloc(vb.buf,vb.used);
   if (!*data_r) *data_r= (char *)vb.buf;
   return adns_s_ok;
